Move test_*.py to tests/core
