+++ .github/workflows/unittest.yml
@@ -86,12 +86,12 @@ jobs:
       if: matrix.python-version != '3.12' && matrix.python-version != '3.13'
       env:
         PYTHONIOENCODING: utf-8
-      run: coverage run -m unittest tests
-    - name: Unit test (core and compact)
+      run: coverage run -m unittest tests.core
+    - name: Unit test (core + compact)
       if: matrix.python-version == '3.12' || matrix.python-version == '3.13'
       env:
         PYTHONIOENCODING: utf-8
-      run: coverage run -m unittest tests tests.compact
+      run: coverage run -m unittest tests.core tests.compact
       # Use 'unittest <test_module>' instead of 'unittest discover' to avoid
       # loading tests with dependencies more than expected.
       # Test cases loaded is defined in __init__.py in the tests directory.
