+++ tests/core/test_cli.py
