+++ tests/core/test_corpus.py
