+++ tests/core/test_khavee.py
