+++ tests/core/test_morpheme.py
