+++ tests/core/test_soundex.py
