+++ tests/core/test_spell.py
