+++ tests/core/test_tag.py
