+++ tests/core/test_tokenize.py
