+++ tests/core/test_tools.py
