+++ tests/core/test_transliterate.py
