+++ tests/core/test_util.py
